+++ routes/images.rb
@@ -1,0 +1,9 @@
+class ELSWC < Sinatra::Application
+
+  get '/images' do
+    @images = Image.all
+    @title = "Welcome to ELSWC"
+    slim :"image/index"
+  end
+ 
+end
