+++ routes/main.rb
@@ -1,0 +1,13 @@
+class ELSWC < Sinatra::Application
+
+  get "/" do
+    @title = "Welcome to ELSWC"
+    slim :index
+  end
+
+  get "/about" do
+    @title = "About ELSWC"
+    slim :about
+  end
+
+end
